+++ server/src/blockchainEventsListener.js
@@ -2,7 +2,6 @@ module.exports = async function (Server) {
   const TOTAL_NUMBER_OF_EVENTS = 7200; // The events to take on the contract to init the db (passed events)
   const MAX_EVENTS_PER_CALL = 200;
   const CALL_NUMBER = Math.ceil(TOTAL_NUMBER_OF_EVENTS / MAX_EVENTS_PER_CALL);
-  const BLOCKS_PROCESSING_INTERVAL = 1000;
 
   // Get the last transactions on the kanvas contract to init the pixels data in db if the server was closed
   let initPixelMap = async function () {
@@ -52,6 +51,9 @@ module.exports = async function (Server) {
       }
     }
   };
+
+  const BLOCKS_PROCESSING_INTERVAL = 1000;
+  Server.volumeInToken = 0; // Init the token volume
 
   let processEvent = async function (event, txId) {
     if (event.source != Server.kanvasContractAddress) return;
@@ -160,6 +162,31 @@ module.exports = async function (Server) {
     //}
   };
 
+  let processOperation = async function (operation) {
+    if (operation?.call_contract?.contract_id != Server.koindxPeripheryAddress)
+      return;
+
+    let decodedOps = await Server.koindxPeripheryContract.decodeOperation(
+      operation
+    );
+
+    const path = decodedOps?.args?.path;
+    if (!path) return;
+
+    const isTokenIn = path[0] == Server.kanvasContractAddress;
+    const isTokenOut = path[path.length - 1] == Server.kanvasContractAddress;
+    if (decodedOps.name == "swap_tokens_out") {
+      if (isTokenIn)
+        Server.volumeInToken += Number(decodedOps.args.amountInMax);
+      else if (isTokenOut)
+        Server.volumeInToken += Number(decodedOps.args.amountOut);
+    } else if (decodedOps.name == "swap_tokens_in") {
+      if (isTokenIn) Server.volumeInToken += Number(decodedOps.args.amountIn);
+      else if (isTokenOut)
+        Server.volumeInToken += Number(decodedOps.args.amountOutMin);
+    }
+  };
+
   // Check for events in the new blocks and process them
   let lastBlockHeight = Number(
     (await Server.client.blockStore.getHighestBlock()).topology.height
@@ -192,9 +219,11 @@ module.exports = async function (Server) {
         lastBlockHeight = newBlockHeight + 1;
       }
 
-      // Process it
+      // Process the block in the right order
       for (let i = blocksToProcess.length - 1; i >= 0; i--) {
         let block = blocksToProcess[i];
+
+        // Process the events
         let transactionReceipts = block.receipt.transaction_receipts;
         let receiptId = block.receipt.id;
         if (transactionReceipts) {
@@ -202,14 +231,28 @@ module.exports = async function (Server) {
             let txReceipt = transactionReceipts[j];
             let events = txReceipt.events;
             if (events) {
-              for (let i = 0; i < events.length; i++) {
+              for (let k = 0; k < events.length; k++) {
                 // Process each event
-                let event = events[i];
+                let event = events[k];
                 await processEvent(event, receiptId);
               }
             }
           }
         }
+
+        // Process the transactions and operations
+        let blockTransactions = block.block.transactions;
+        if (blockTransactions) {
+          for (let j = 0; j < blockTransactions.length; j++) {
+            let transaction = blockTransactions[j];
+            let txOperations = transaction.operations;
+            for (let k = 0; k < txOperations.length; k++) {
+              let txOperation = txOperations[k];
+              await processOperation(txOperation);
+            }
+          }
+        }
+
         Server.infoLogging(
           "Processed block",
           block.block_height,
@@ -220,6 +263,9 @@ module.exports = async function (Server) {
       Server.errorLogging("Error while retrieving last block", err);
     }
   }, BLOCKS_PROCESSING_INTERVAL);
+
+  // Uncomment to init from blockchain state, don't forget to modify the TOTAL_NUMBER_OF_EVENTS
+  //await initPixelMap();
 
   // Old processing function
   /*void (async () => {
@@ -243,27 +289,17 @@ module.exports = async function (Server) {
     }
   })();*/
 
-  // Uncomment to init from blockchain state, don't forget to modify the TOTAL_NUMBER_OF_EVENTS
-  //await initPixelMap();
-
-  /*let block = await Server.client.blockStore.getBlocksById([
-    "0x12201a4b7c004ce7c9fe4836951599dc290cd430efe68f134f36df2c5318d8adf5e7",
-  ]);
   /*let block = await Server.client.blockStore.getBlocksByHeight(
     "0x12208bc173edc126acee50f34f3a962a5dc46f6d18842b2629604e4f33b0a9afde8d",
     8186122,
     1
   );
-  block = block.block_items[0];
-  let blockOperations = block.block.transactions[0].operations;
-  blockOperations.forEach((ops) => {
-    let isCallContract = !!ops.call_contract;
-    if (isCallContract) {
-      if (ops.call_contract.contract_id == Server.koindxCoreAddress) {
-        console.log(ops.call_contract.args);
-      }
-    }
-  });*/
+*/
+
+  /*let block = await Server.client.blockStore.getBlocksById([
+    "0x12201a4b7c004ce7c9fe4836951599dc290cd430efe68f134f36df2c5318d8adf5e7",
+  ]);
+  block = block.block_items[0];*/
 
   /*let transactionReceipts = block.receipt.transaction_receipts;
   let receiptId = block.receipt.id;